arquitetura_de_dados/pyproject.toml
[project]
name = "arquitetura_de_dados"
version = "0.0.1"
authors = [{ name = "fabio.jdluz@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "databricks-dlt",
    "databricks-feature-engineering",
    "databricks-connect>=14.1",
    "ipykernel",
]

[project.scripts]
main = "app2.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.hatch.build.targets.wheel]
packages = ["src"]
